# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Medical 3D Shared Viewer API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /shared-viewers/{token}/manifest:
    get:
      operationId: getSharedViewerManifest
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
            minLength: 32
            maxLength: 256
      responses:
        '200':
          description: Viewer manifest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ViewerManifest'
        '404':
          description: Unknown share
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '410':
          description: Expired or revoked share
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
  /shared-viewers/{token}/plans:
    post:
      operationId: createViewerPlan
      parameters:
        - $ref: '#/components/parameters/ShareToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SavePlanRequest' }
      responses:
        '201': { description: Plan created }
        '403': { description: Save permission denied }
        '422': { description: Invalid scene state }
  /shared-viewers/{token}/plans/{planId}:
    get:
      operationId: getViewerPlan
      parameters:
        - $ref: '#/components/parameters/ShareToken'
        - $ref: '#/components/parameters/PlanId'
      responses:
        '200': { description: Viewer plan }
        '404': { description: Plan not found }
    put:
      operationId: updateViewerPlan
      parameters:
        - $ref: '#/components/parameters/ShareToken'
        - $ref: '#/components/parameters/PlanId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SavePlanRequest' }
      responses:
        '200': { description: Plan updated }
        '409': { description: Plan version conflict }
        '422': { description: Invalid scene state }
    delete:
      operationId: deleteViewerPlan
      parameters:
        - $ref: '#/components/parameters/ShareToken'
        - $ref: '#/components/parameters/PlanId'
      responses:
        '204': { description: Plan deleted }
components:
  parameters:
    ShareToken:
      in: path
      name: token
      required: true
      schema: { type: string, minLength: 32, maxLength: 256 }
    PlanId:
      in: path
      name: planId
      required: true
      schema: { type: string, format: uuid }
  schemas:
    ViewerManifest:
      type: object
      required: [resultId, title, unit, coordinateSystem, manifestVersion, permissions, layers]
      properties:
        resultId: { type: string, format: uuid }
        title: { type: string }
        unit: { type: string, enum: [mm] }
        coordinateSystem: { type: string, enum: [LPS, RAS] }
        manifestVersion: { type: integer, minimum: 1 }
        permissions: { $ref: '#/components/schemas/ViewerPermissions' }
        layers:
          type: array
          items: { $ref: '#/components/schemas/ModelLayer' }
    ViewerPermissions:
      type: object
      required: [view, measure, annotate, savePlan, download]
      properties:
        view: { type: boolean }
        measure: { type: boolean }
        annotate: { type: boolean }
        savePlan: { type: boolean }
        download: { type: boolean }
    ModelLayer:
      type: object
      required: [id, code, name, color, opacity, visible, sortOrder, assets]
      properties:
        id: { type: string, format: uuid }
        parentId: { type: [string, 'null'], format: uuid }
        code: { type: string }
        name: { type: string }
        color: { type: string, pattern: '^#[0-9A-Fa-f]{6}([0-9A-Fa-f]{2})?$' }
        opacity: { type: number, minimum: 0, maximum: 1 }
        visible: { type: boolean }
        volumeMl: { type: [number, 'null'], minimum: 0 }
        sortOrder: { type: integer }
        assets: { $ref: '#/components/schemas/AssetVariants' }
    AssetVariants:
      type: object
      additionalProperties: false
      properties:
        low: { type: string, format: uri }
        medium: { type: string, format: uri }
        high: { type: string, format: uri }
        canonical: { type: string, format: uri }
    ApiError:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }
        details: { type: object, additionalProperties: true }
    SavePlanRequest:
      type: object
      required: [name, expectedVersion, sceneState, measurements, annotations]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        expectedVersion: { type: integer, minimum: 0 }
        sceneState: { type: object, additionalProperties: true }
        measurements:
          type: array
          maxItems: 200
          items: { $ref: '#/components/schemas/MeasurementEvent' }
        annotations:
          type: array
          maxItems: 200
          items: { $ref: '#/components/schemas/AnnotationEvent' }
    Point3D:
      type: object
      required: [x, y, z]
      properties:
        x: { type: number }
        y: { type: number }
        z: { type: number }
    MeasurementEvent:
      type: object
      required: [id, type, points, style]
      properties:
        id: { type: string, format: uuid }
        layerId: { type: [string, 'null'], format: uuid }
        type: { type: string, enum: [LENGTH, DIAMETER, ANGLE, CLOSED_AREA] }
        points:
          type: array
          minItems: 1
          maxItems: 128
          items: { $ref: '#/components/schemas/Point3D' }
        style: { type: object, additionalProperties: true }
    AnnotationEvent:
      type: object
      required: [id, text, point, style]
      properties:
        id: { type: string, format: uuid }
        layerId: { type: [string, 'null'], format: uuid }
        text: { type: string, minLength: 1, maxLength: 1000 }
        point: { $ref: '#/components/schemas/Point3D' }
        style: { type: object, additionalProperties: true }